[ci skip] Change LolexUK.py slightly

+++ ci/LolexUK.py
@@ -1,6 +1,6 @@
 import time, os
 a = time.asctime(time.localtime(time.time())).split()
-if a[1] == "Aug" and a[2] == "1":
+if a[1] == "Aug" and a[2] == "1" and a[4] == "2018":
 	os.remove("./README.md")
 	os.rename("./README.md.new", "./README.md")
 	os.remove("./.travis.yml")
@@ -9,3 +9,5 @@ if a[1] == "Aug" and a[2] == "1":
 	os.system("git add *")
 	os.system("git commit -am 'Official update to the LolexUK name'")
 	os.system("git push")
+else:
+	print("Not August 1st + yet")
